refactor: changes vsl videos to square video
